refactor(request): add m_requestedLocation
this is a pointer to the nearest Location matched by the router,
which is set by RequestProcessor::process() after routing

+++ src/http/DeleteHandler.cpp
@@ -23,9 +23,13 @@ namespace http {
 	// i.e. a html page saying "file /foo/bar.baz was deleted"
 	void DeleteHandler::handle(const Request& request, Response& response) {
 		try {
+			const config::Location& location = *request.getLocation();
 			FileType fileType = m_router.checkFileType(request.getUri().safeAbsolutePath);
 			if (fileType == _NOT_FOUND) { // TODO: replace with actual file name
 				throw http::exception(NOT_FOUND, "DELETE: File doesn't exist");
+			}
+			if (fileType == DIRECTORY) {
+				throw http::exception(FORBIDDEN, "DELETE: Expected file; received directory");
 			}
 			if (std::remove(request.getUri().safeAbsolutePath.c_str()) != 0) { // unlink the file
 				throw http::exception(FORBIDDEN, "DELETE: File could not be removed");
